Fixing GUI
	deleted:    lib/j3dcore.jar
	deleted:    lib/j3dutils.jar
	modified:   src/mosaic/ia/Analysis.java
	modified:   src/mosaic/ia/gui/GUIDesign.java
	modified:   src/mosaic/plugins/Calibration_.java

+++ plugin/src/mosaic/plugins/Calibration_.java
@@ -42,7 +42,7 @@ public class Calibration_ implements PlugIn, PreviewInterface {
 		allocateTwoImages();
 		detect();
 		// changed by arun to affect janick's refactoring
-		new ParticleLinker().linkParticles(frames, frames_number, linkrange, displacement);
+		new ParticleLinker().linkParticles(frames, frames_number, linkrange, (float)displacement);
 		//
 		calculateShifts();
 		regression();
